+++ apps/doop/src/lib/store/createGlobalsSlice.tsx
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Greenhouse contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+const createGlobalsSlice = (set: any) => ({
+  globals: {
+    endpoint: "",
+
+    actions: {
+      setEndpoint: (endpoint: any) =>
+        set(
+          (state: any) => ({
+            globals: { ...state.globals, endpoint: endpoint },
+          }),
+          false,
+          "globals/setEndpoint"
+        ),
+    },
+  },
+})
+
+export default createGlobalsSlice